+++ src/mdschema/mod.rs
@@ -1,7 +1,4 @@
-pub mod errors;
 pub mod reports;
 pub mod validator;
 
-pub use errors::{ErrorSeverity, ValidatorError};
-pub use reports::ValidatorReport;
 pub use validator::zipper_tree_validator::ValidationZipperTree;
