+++ src/mdschema/validator/validator.rs
@@ -1,0 +1,10 @@
+use crate::mdschema::reports::validation_report::ValidatorReport;
+
+pub trait Validator {
+    fn new(schema_str: &str, input_str: &str) -> Result<Self, Box<dyn std::error::Error>>
+    where
+        Self: Sized;
+    fn validate(&mut self) -> Result<(), Box<dyn std::error::Error>>;
+    fn read_input(&mut self, input: &str) -> Result<(), Box<dyn std::error::Error>>;
+    fn report(&self) -> ValidatorReport;
+}
